Update README.rst
Enhanced the 'Parse URL' section by providing additional details about the protocol and resource components of the URL. Included explanations for the protocol 'http' and its expansion, as well as clarification on the resource '/' indicating the retrieval of the main (index) page. This addition aims to offer a more comprehensive understanding of the information contained within a URL. Thank you for considering this enhancement!